+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Expose the farmOS instance name, version, and system of measurement
+
 ## [1.0.0-alpha13] - 2024-04-14
 
 ### Added